lib/fargate/requirements.txt
boto3
fastapi
uvicorn
c2pa-python